[betty] create other file contains a function
