+++ general.c
@@ -156,14 +156,3 @@ int myexec(char *line_ptr, char *exec_path, char *args[], char *env_args[])
 	printf("el status es: %d\n", wstatus);
 	return (wstatus);
 }
-/**
- *not_found_command - print comand not found
- *@argv_0: command not found
- */
-void not_found_command(int count, char *argv_0)
-{
-	write(STDOUT_FILENO, argv_0, _strlen(argv_0));
-	write(STDOUT_FILENO, ": ", 2);
-	print_number(count);
-	write(STDOUT_FILENO, ": command not found\n", 20);
-}
